Fix TimesNet bf16 incompatibility + add TN-specific batch=128 sbatch
models/timesnet.py:
- torch.fft.rfft does not support BFloat16/Float16. Cast x.float() inside
  FFT_for_Period so the FFT runs in fp32 even under autocast(bf16). The
  result is only used to pick top-k period indices (no gradient flow), so
  the cast is correct and cheap.
- Symptom before fix: all 10 first-attempt TimesNet jobs crashed at
  models/timesnet.py:8 with 'RuntimeError: Unsupported dtype BFloat16'
  on stock 0 of Round 1.

scripts/TN_seq.sbatch, scripts/TN_glob.sbatch:
- TimesNet's Inception_Block_V1 + FFT-derived period reshapes blow up
  intermediate tensor memory at short horizons (H=5 sequential and global
  both OOM'd at batch=512 on A100 80GB during loss.backward).
- New TN-specific sbatch templates use batch_size=128 (4x headroom).
  H=60+ sequential and H=20+ global survive at 512, so the iT_*.sbatch
  templates are still fine for the longer horizons that finished first.

+++ models/timesnet.py
@@ -5,7 +5,10 @@ from layers.Embed import DataEmbedding
 from layers.Conv_Blocks import Inception_Block_V1
 
 def FFT_for_Period(x, k=2):
-    xf = torch.fft.rfft(x, dim=1)
+    # torch.fft.rfft does not support BFloat16/Float16 — cast to fp32 for the
+    # transform under AMP autocast. The result is real-magnitude only used to
+    # pick top-k periods (indices), so dtype of xf does not flow into grads.
+    xf = torch.fft.rfft(x.float(), dim=1)
     frequency_list = abs(xf).mean(0).mean(-1)
     frequency_list[0] = 0
     _, top_list = torch.topk(frequency_list, k)
